requirements.txt
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
matplotlib==3.8.4
pillow==10.3.0
python-dotenv==1.0.1